Various fixes and logging improvements
* Cast to long in the helper not integer to make sure we actually get 10 buckets
* Add some logging and change log levels
* Split out the SM and E detection into a separate method to make the looping method cleaner

+++ src/main/groovy/uk/ac/ox/softeng/maurodatamapper/plugins/database/summarymetadata/LongIntervalHelper.groovy
@@ -68,7 +68,7 @@ class LongIntervalHelper extends AbstractIntervalHelper<Long> {
             intervalLength = 10000000
         } else {
             Double base = Math.log10((Double) ((maxValue - minValue)  / 10))
-            intervalLength = (Integer) Math.pow(10, Math.ceil(base))
+            intervalLength = (Long) Math.pow(10, Math.ceil(base))
         }
 
         firstIntervalStart = minValue.intdiv(intervalLength) * intervalLength
